+++ packages/engine/Source/Scene/Empty3DTileContent.js
@@ -150,6 +150,10 @@ Empty3DTileContent.prototype.applyStyle = function (style) {};
 
 Empty3DTileContent.prototype.update = function (tileset, frameState) {};
 
+Empty3DTileContent.prototype.pick = function (ray, frameState, result) {
+  return undefined;
+};
+
 Empty3DTileContent.prototype.isDestroyed = function () {
   return false;
 };