+++ packages/engine/Source/Shaders/Model/AtmosphereStageVS.glsl
@@ -1,0 +1,12 @@
+void atmosphereStage(ProcessedAttributes attributes) {
+    vec3 lightDirection = czm_getDynamicAtmosphereLightDirection(v_positionWC, czm_atmosphereDynamicLighting);
+
+    czm_computeGroundAtmosphereScattering(
+        // This assumes the geometry stage came before this.
+        v_positionWC,
+        lightDirection,
+        v_atmosphereRayleighColor,
+        v_atmosphereMieColor,
+        v_atmosphereOpacity
+    );
+}