+++ packages/engine/Specs/Scene/Model/ModelSpec.js
@@ -1,4 +1,5 @@
 import {
+  Atmosphere,
   Axis,
   Cartesian2,
   Cartesian3,
@@ -10,13 +11,17 @@ import {
   Color,
   ColorBlendMode,
   Credit,
+  CustomShader,
   defaultValue,
   defined,
+  DirectionalLight,
   DistanceDisplayCondition,
+  DynamicAtmosphereLightingType,
   DracoLoader,
   Ellipsoid,
   Event,
   FeatureDetection,
+  Fog,
   HeadingPitchRange,
   HeadingPitchRoll,
   HeightReference,
@@ -39,6 +44,7 @@ import {
   ShadowMode,
   SplitDirection,
   StyleCommandsNeeded,
+  SunLight,
   Transforms,
   WireframeIndexGenerator,
 } from "../../../index.js";
@@ -104,6 +110,11 @@ describe(
       "./Data/Models/glTF-2.0/CesiumMan/glTF-Draco/CesiumMan.gltf";
     const boxCesiumRtcUrl =
       "./Data/Models/glTF-2.0/BoxCesiumRtc/glTF/BoxCesiumRtc.gltf";
+
+    const propertyTextureWithTextureTransformUrl =
+      "./Data/Models/glTF-2.0/PropertyTextureWithTextureTransform/glTF/PropertyTextureWithTextureTransform.gltf";
+    const featureIdTextureWithTextureTransformUrl =
+      "./Data/Models/glTF-2.0/FeatureIdTextureWithTextureTransform/glTF/FeatureIdTextureWithTextureTransform.gltf";
 
     const fixedFrameTransform = Transforms.localFrameToFixedFrameGenerator(
       "north",
@@ -712,6 +723,141 @@ describe(
         ).then(function (model) {
           verifyRender(model, true);
         });
+      });
+    });
+
+    it("transforms property textures with KHR_texture_transform", async function () {
+      const resource = Resource.createIfNeeded(
+        propertyTextureWithTextureTransformUrl
+      );
+      // The texture in the example model contains contains 8x8 pixels
+      // with increasing 'red' component values [0 to 64)*3, interpreted
+      // as a normalized `UINT8` property.
+      // It has a transform with an offset of [0.25, 0.25], and a scale
+      // of [0.5, 0.5].
+      // Create a custom shader that will render any value that is smaller
+      // than 16*3 or larger than 48*3 (i.e. the first two rows of pixels
+      // or the last two rows of pixels) as completely red.
+      // These pixels should NOT be visible when the transform is applied.
+      const customShader = new CustomShader({
+        fragmentShaderText: `
+          void fragmentMain(FragmentInput fsInput, inout czm_modelMaterial material)
+          {
+            float value = float(fsInput.metadata.exampleProperty);
+            float i = value * 255.0;
+            if (i < 16.0 * 3.0) {
+              material.diffuse = vec3(1.0, 0.0, 0.0);
+            } else if (i >= 48.0 * 3.0) {
+              material.diffuse = vec3(1.0, 0.0, 0.0);
+            } else {
+              material.diffuse = vec3(0.0, 0.0, 0.0);
+            }
+          }
+          `,
+      });
+
+      const gltf = await resource.fetchJson();
+      await loadAndZoomToModelAsync(
+        {
+          gltf: gltf,
+          basePath: propertyTextureWithTextureTransformUrl,
+          customShader: customShader,
+          // This is important to make sure that the property
+          // texture is fully loaded when the model is rendered!
+          incrementallyLoadTextures: false,
+        },
+        scene
+      );
+      const renderOptions = {
+        scene: scene,
+        time: defaultDate,
+      };
+      // Move the camera to look at the point (0.1, 0.1) of
+      // the plane at a distance of 0.15. (Note that the axes
+      // are swapped, apparently - 'x' is the distance)
+      scene.camera.position = new Cartesian3(0.15, 0.1, 0.1);
+      scene.camera.direction = Cartesian3.negate(
+        Cartesian3.UNIT_X,
+        new Cartesian3()
+      );
+      scene.camera.up = Cartesian3.clone(Cartesian3.UNIT_Z);
+      scene.camera.frustum.near = 0.01;
+      scene.camera.frustum.far = 5.0;
+
+      // When the texture transform was applied, then the
+      // resulting pixels should be nearly black (or at
+      // least not red)
+      expect(renderOptions).toRenderAndCall(function (rgba) {
+        expect(rgba[0]).toBeLessThan(50);
+        expect(rgba[1]).toBeLessThan(50);
+        expect(rgba[2]).toBeLessThan(50);
+      });
+    });
+
+    it("transforms feature ID textures with KHR_texture_transform", async function () {
+      const resource = Resource.createIfNeeded(
+        featureIdTextureWithTextureTransformUrl
+      );
+      // The texture in the example model contains contains 8x8 pixels
+      // with increasing 'red' component values [0 to 64)*3.
+      // It has a transform with an offset of [0.25, 0.25], and a scale
+      // of [0.5, 0.5].
+      // Create a custom shader that will render any value that is smaller
+      // than 16*3 or larger than 48*3 (i.e. the first two rows of pixels
+      // or the last two rows of pixels) as completely red.
+      // These pixels should NOT be visible when the transform is applied.
+      const customShader = new CustomShader({
+        fragmentShaderText: `
+        void fragmentMain(FragmentInput fsInput, inout czm_modelMaterial material)
+        {
+          int id = fsInput.featureIds.featureId_0;
+          if (id < 16 * 3) {
+            material.diffuse = vec3(1.0, 0.0, 0.0);
+          } else if (id >= 48 * 3) {
+            material.diffuse = vec3(1.0, 0.0, 0.0);
+          } else {
+            material.diffuse = vec3(0.0, 0.0, 0.0);
+          }
+        }
+      `,
+      });
+
+      const gltf = await resource.fetchJson();
+      await loadAndZoomToModelAsync(
+        {
+          gltf: gltf,
+          basePath: featureIdTextureWithTextureTransformUrl,
+          customShader: customShader,
+          // This is important to make sure that the feature ID
+          // texture is fully loaded when the model is rendered!
+          incrementallyLoadTextures: false,
+        },
+        scene
+      );
+      const renderOptions = {
+        scene: scene,
+        time: defaultDate,
+      };
+      // Move the camera to look at the point (0.1, 0.1) of
+      // the plane at a distance of 0.15. (Note that the axes
+      // are swapped, apparently - 'x' is the distance)
+      //scene.camera.position = new Cartesian3(0.15, 0.1, 0.1);
+      scene.camera.position = new Cartesian3(0.15, 0.1, 0.1);
+      scene.camera.direction = Cartesian3.negate(
+        Cartesian3.UNIT_X,
+        new Cartesian3()
+      );
+      scene.camera.up = Cartesian3.clone(Cartesian3.UNIT_Z);
+      scene.camera.frustum.near = 0.01;
+      scene.camera.frustum.far = 5.0;
+
+      // When the texture transform was applied, then the
+      // resulting pixels should be nearly black (or at
+      // least not red)
+      expect(renderOptions).toRenderAndCall(function (rgba) {
+        expect(rgba[0]).toBeLessThan(50);
+        expect(rgba[1]).toBeLessThan(50);
+        expect(rgba[2]).toBeLessThan(50);
       });
     });
 
@@ -4409,6 +4555,463 @@ describe(
           verifyRender(model, false);
         });
       });
+    });
+
+    describe("fog", function () {
+      const sunnyDate = JulianDate.fromIso8601("2024-01-11T15:00:00Z");
+      const darkDate = JulianDate.fromIso8601("2024-01-11T00:00:00Z");
+
+      afterEach(function () {
+        scene.atmosphere = new Atmosphere();
+        scene.fog = new Fog();
+        scene.light = new SunLight();
+        scene.camera.switchToPerspectiveFrustum();
+      });
+
+      function viewFog(scene, model) {
+        // In order for fog to create a visible change, the camera needs to be
+        // further away from the model. This would make the box sub-pixel
+        // so to make it fill the canvas, use an ortho camera the same
+        // width of the box to make the scene look 2D.
+        const center = model.boundingSphere.center;
+        scene.camera.lookAt(center, new Cartesian3(1000, 0, 0));
+        scene.camera.switchToOrthographicFrustum();
+        scene.camera.frustum.width = 1;
+      }
+
+      it("renders a model in fog", async function () {
+        // Move the fog very close to the camera;
+        scene.fog.density = 1.0;
+
+        // Increase the brightness to make the fog color
+        // stand out more for this test
+        scene.atmosphere.brightnessShift = 1.0;
+
+        const model = await loadAndZoomToModelAsync(
+          {
+            url: boxTexturedGltfUrl,
+            modelMatrix: Transforms.eastNorthUpToFixedFrame(
+              Cartesian3.fromDegrees(0, 0, 10.0)
+            ),
+          },
+          scene
+        );
+
+        viewFog(scene, model);
+
+        const renderOptions = {
+          scene,
+          time: sunnyDate,
+        };
+
+        // First, turn off the fog to capture the original color
+        let originalColor;
+        scene.fog.enabled = false;
+        expect(renderOptions).toRenderAndCall(function (rgba) {
+          originalColor = rgba;
+          expect(originalColor).not.toEqual([0, 0, 0, 255]);
+        });
+
+        // Now turn on fog. The result should be bluish
+        // than before due to scattering.
+        scene.fog.enabled = true;
+        expect(renderOptions).toRenderAndCall(function (rgba) {
+          expect(rgba).not.toEqual([0, 0, 0, 255]);
+          expect(rgba).not.toEqual(originalColor);
+
+          // The result should have a bluish tint
+          const [r, g, b, a] = rgba;
+          expect(b).toBeGreaterThan(r);
+          expect(b).toBeGreaterThan(g);
+          expect(a).toBe(255);
+        });
+      });
+
+      it("renders a model in fog (sunlight)", async function () {
+        // Move the fog very close to the camera;
+        scene.fog.density = 1.0;
+
+        // Increase the brightness to make the fog color
+        // stand out more for this test
+        scene.atmosphere.brightnessShift = 1.0;
+
+        const model = await loadAndZoomToModelAsync(
+          {
+            url: boxTexturedGltfUrl,
+            modelMatrix: Transforms.eastNorthUpToFixedFrame(
+              Cartesian3.fromDegrees(0, 0, 10.0)
+            ),
+          },
+          scene
+        );
+
+        // In order for fog to render, the camera needs to be
+        // further away from the model. This would make the box sub-pixel
+        // so to make it fill the canvas, use an ortho camera the same
+        // width of the box to make the scene look 2D.
+        const center = model.boundingSphere.center;
+        scene.camera.lookAt(center, new Cartesian3(1000, 0, 0));
+        scene.camera.switchToOrthographicFrustum();
+        scene.camera.frustum.width = 1;
+
+        // Grab the color when dynamic lighting is off for comparison
+        scene.atmosphere.dynamicLighting = DynamicAtmosphereLightingType.NONE;
+        const renderOptions = {
+          scene,
+          time: sunnyDate,
+        };
+        let originalColor;
+        expect(renderOptions).toRenderAndCall(function (rgba) {
+          originalColor = rgba;
+          expect(originalColor).not.toEqual([0, 0, 0, 255]);
+        });
+
+        // switch the lighting model to sunlight
+        scene.atmosphere.dynamicLighting =
+          DynamicAtmosphereLightingType.SUNLIGHT;
+
+        // Render in the sun, it should be a different color than the
+        // original
+        let sunnyColor;
+        expect(renderOptions).toRenderAndCall(function (rgba) {
+          sunnyColor = rgba;
+          expect(rgba).not.toEqual([0, 0, 0, 255]);
+          expect(rgba).not.toEqual(originalColor);
+        });
+
+        // Render in the dark, it should be a different color and
+        // darker than in sun
+        renderOptions.time = darkDate;
+        expect(renderOptions).toRenderAndCall(function (rgba) {
+          expect(rgba).not.toEqual([0, 0, 0, 255]);
+          expect(rgba).not.toEqual(originalColor);
+          expect(rgba).not.toEqual(sunnyColor);
+
+          const [sunR, sunG, sunB, sunA] = sunnyColor;
+          const [r, g, b, a] = rgba;
+          expect(r).toBeLessThan(sunR);
+          expect(g).toBeLessThan(sunG);
+          expect(b).toBeLessThan(sunB);
+          expect(a).toBe(sunA);
+        });
+      });
+
+      it("renders a model in fog (scene light)", async function () {
+        // Move the fog very close to the camera;
+        scene.fog.density = 1.0;
+
+        // Increase the brightness to make the fog color
+        // stand out more for this test
+        scene.atmosphere.brightnessShift = 1.0;
+
+        const model = await loadAndZoomToModelAsync(
+          {
+            url: boxTexturedGltfUrl,
+            modelMatrix: Transforms.eastNorthUpToFixedFrame(
+              Cartesian3.fromDegrees(0, 0, 10.0)
+            ),
+          },
+          scene
+        );
+
+        viewFog(scene, model);
+
+        // Grab the color when dynamic lighting is off for comparison
+        scene.atmosphere.dynamicLighting = DynamicAtmosphereLightingType.NONE;
+        const renderOptions = {
+          scene,
+          time: sunnyDate,
+        };
+        let originalColor;
+        expect(renderOptions).toRenderAndCall(function (rgba) {
+          originalColor = rgba;
+          expect(originalColor).not.toEqual([0, 0, 0, 255]);
+        });
+
+        // Also grab the color in sunlight for comparison
+        scene.atmosphere.dynamicLighting =
+          DynamicAtmosphereLightingType.SUNLIGHT;
+        let sunnyColor;
+        expect(renderOptions).toRenderAndCall(function (rgba) {
+          sunnyColor = rgba;
+          expect(rgba).not.toEqual([0, 0, 0, 255]);
+          expect(rgba).not.toEqual(originalColor);
+        });
+
+        // Set a light on the scene, but since dynamicLighting is SUNLIGHT,
+        // it should have no effect yet
+        scene.light = new DirectionalLight({
+          direction: new Cartesian3(0, 1, 0),
+        });
+        expect(renderOptions).toRenderAndCall(function (rgba) {
+          expect(rgba).toEqual(sunnyColor);
+        });
+
+        // Set dynamic lighting to use the scene light, now it should
+        // render a different color from the other light sources
+        scene.atmosphere.dynamicLighting =
+          DynamicAtmosphereLightingType.SCENE_LIGHT;
+
+        expect(renderOptions).toRenderAndCall(function (rgba) {
+          expect(rgba).not.toEqual([0, 0, 0, 255]);
+          expect(rgba).not.toEqual(originalColor);
+          expect(rgba).not.toEqual(sunnyColor);
+        });
+      });
+
+      it("adjusts atmosphere light intensity", async function () {
+        // Move the fog very close to the camera;
+        scene.fog.density = 1.0;
+
+        // Increase the brightness to make the fog color
+        // stand out more. We'll use the light intensity to
+        // modulate this.
+        scene.atmosphere.brightnessShift = 1.0;
+
+        const model = await loadAndZoomToModelAsync(
+          {
+            url: boxTexturedGltfUrl,
+            modelMatrix: Transforms.eastNorthUpToFixedFrame(
+              Cartesian3.fromDegrees(0, 0, 10.0)
+            ),
+          },
+          scene
+        );
+
+        viewFog(scene, model);
+
+        const renderOptions = {
+          scene,
+          time: sunnyDate,
+        };
+
+        // Grab the original color.
+        let originalColor;
+        expect(renderOptions).toRenderAndCall(function (rgba) {
+          originalColor = rgba;
+          expect(rgba).not.toEqual([0, 0, 0, 255]);
+
+          // The result should have a bluish tint from the atmosphere
+          const [r, g, b, a] = rgba;
+          expect(b).toBeGreaterThan(r);
+          expect(b).toBeGreaterThan(g);
+          expect(a).toBe(255);
+        });
+
+        // Turn down the light intensity
+        scene.atmosphere.lightIntensity = 5.0;
+        expect(renderOptions).toRenderAndCall(function (rgba) {
+          expect(rgba).not.toEqual([0, 0, 0, 255]);
+          expect(rgba).not.toEqual(originalColor);
+
+          // Check that each component (except alpha) is darker than before
+          const [oldR, oldG, oldB, oldA] = originalColor;
+          const [r, g, b, a] = rgba;
+          expect(r).toBeLessThan(oldR);
+          expect(g).toBeLessThan(oldG);
+          expect(b).toBeLessThan(oldB);
+          expect(a).toBe(oldA);
+        });
+      });
+
+      it("applies a hue shift", async function () {
+        // Move the fog very close to the camera;
+        scene.fog.density = 1.0;
+
+        // Increase the brightness to make the fog color
+        // stand out more for this test
+        scene.atmosphere.brightnessShift = 1.0;
+
+        const model = await loadAndZoomToModelAsync(
+          {
+            url: boxTexturedGltfUrl,
+            modelMatrix: Transforms.eastNorthUpToFixedFrame(
+              Cartesian3.fromDegrees(0, 0, 10.0)
+            ),
+          },
+          scene
+        );
+
+        viewFog(scene, model);
+
+        const renderOptions = {
+          scene,
+          time: sunnyDate,
+        };
+
+        // Grab the original color.
+        let originalColor;
+        expect(renderOptions).toRenderAndCall(function (rgba) {
+          originalColor = rgba;
+          expect(rgba).not.toEqual([0, 0, 0, 255]);
+
+          // The result should have a bluish tint from the atmosphere
+          const [r, g, b, a] = rgba;
+          expect(b).toBeGreaterThan(r);
+          expect(b).toBeGreaterThan(g);
+          expect(a).toBe(255);
+        });
+
+        // Shift the fog color to be reddish
+        scene.atmosphere.hueShift = 0.4;
+        let redColor;
+        expect(renderOptions).toRenderAndCall(function (rgba) {
+          redColor = rgba;
+          expect(rgba).not.toEqual([0, 0, 0, 255]);
+          expect(redColor).not.toEqual(originalColor);
+
+          // Check for a reddish tint
+          const [r, g, b, a] = rgba;
+          expect(r).toBeGreaterThan(g);
+          expect(r).toBeGreaterThan(b);
+          expect(a).toBe(255);
+        });
+
+        // ...now greenish
+        scene.atmosphere.hueShift = 0.7;
+        let greenColor;
+        expect(renderOptions).toRenderAndCall(function (rgba) {
+          greenColor = rgba;
+          expect(rgba).not.toEqual([0, 0, 0, 255]);
+          expect(greenColor).not.toEqual(originalColor);
+          expect(greenColor).not.toEqual(redColor);
+
+          // Check for a greenish tint
+          const [r, g, b, a] = rgba;
+          expect(g).toBeGreaterThan(r);
+          expect(g).toBeGreaterThan(b);
+          expect(a).toBe(255);
+        });
+
+        // ...and all the way around the color wheel back to bluish
+        scene.atmosphere.hueShift = 1.0;
+        expect(renderOptions).toRenderAndCall(function (rgba) {
+          expect(rgba).not.toEqual([0, 0, 0, 255]);
+          expect(rgba).toEqual(originalColor);
+        });
+      });
+
+      it("applies a brightness shift", async function () {
+        // Move the fog very close to the camera;
+        scene.fog.density = 1.0;
+
+        const model = await loadAndZoomToModelAsync(
+          {
+            url: boxTexturedGltfUrl,
+            modelMatrix: Transforms.eastNorthUpToFixedFrame(
+              Cartesian3.fromDegrees(0, 0, 10.0)
+            ),
+          },
+          scene
+        );
+
+        viewFog(scene, model);
+
+        const renderOptions = {
+          scene,
+          time: sunnyDate,
+        };
+
+        // Grab the original color.
+        let originalColor;
+        scene.atmosphere.brightnessShift = 1.0;
+        expect(renderOptions).toRenderAndCall(function (rgba) {
+          originalColor = rgba;
+          expect(rgba).not.toEqual([0, 0, 0, 255]);
+
+          // The result should have a bluish tint from the atmosphere
+          const [r, g, b, a] = rgba;
+          expect(b).toBeGreaterThan(r);
+          expect(b).toBeGreaterThan(g);
+          expect(a).toBe(255);
+        });
+
+        // Turn down the brightness
+        scene.atmosphere.brightnessShift = 0.5;
+        expect(renderOptions).toRenderAndCall(function (rgba) {
+          expect(rgba).not.toEqual([0, 0, 0, 255]);
+          expect(rgba).not.toEqual(originalColor);
+
+          // Check that each component (except alpha) is darker than before
+          const [oldR, oldG, oldB, oldA] = originalColor;
+          const [r, g, b, a] = rgba;
+          expect(r).toBeLessThan(oldR);
+          expect(g).toBeLessThan(oldG);
+          expect(b).toBeLessThan(oldB);
+          expect(a).toBe(oldA);
+        });
+      });
+
+      it("applies a saturation shift", async function () {
+        // Move the fog very close to the camera;
+        scene.fog.density = 1.0;
+
+        const model = await loadAndZoomToModelAsync(
+          {
+            url: boxTexturedGltfUrl,
+            modelMatrix: Transforms.eastNorthUpToFixedFrame(
+              Cartesian3.fromDegrees(0, 0, 10.0)
+            ),
+          },
+          scene
+        );
+
+        viewFog(scene, model);
+
+        const renderOptions = {
+          scene,
+          time: sunnyDate,
+        };
+
+        // Grab the original color.
+        let originalColor;
+        expect(renderOptions).toRenderAndCall(function (rgba) {
+          originalColor = rgba;
+          expect(rgba).not.toEqual([0, 0, 0, 255]);
+
+          // The result should have a bluish tint from the atmosphere
+          const [r, g, b, a] = rgba;
+          expect(b).toBeGreaterThan(r);
+          expect(b).toBeGreaterThan(g);
+          expect(a).toBe(255);
+        });
+
+        // Turn down the saturation all the way
+        scene.atmosphere.saturationShift = -1.0;
+        expect(renderOptions).toRenderAndCall(function (rgba) {
+          expect(rgba).not.toEqual([0, 0, 0, 255]);
+          expect(rgba).not.toEqual(originalColor);
+
+          // Check that each component (except alpha) is the same
+          // as grey values have R = G = B
+          const [r, g, b, a] = rgba;
+          expect(g).toBe(r);
+          expect(b).toBe(g);
+          expect(a).toBe(255);
+        });
+      });
+    });
+
+    it("pick returns position of intersection between ray and model surface", async function () {
+      const model = await loadAndZoomToModelAsync(
+        {
+          url: boxTexturedGltfUrl,
+          enablePick: !scene.frameState.context.webgl2,
+        },
+        scene
+      );
+      const ray = scene.camera.getPickRay(
+        new Cartesian2(
+          scene.drawingBufferWidth / 2.0,
+          scene.drawingBufferHeight / 2.0
+        )
+      );
+
+      const expected = new Cartesian3(0.5, 0, 0.5);
+      expect(model.pick(ray, scene.frameState)).toEqualEpsilon(
+        expected,
+        CesiumMath.EPSILON12
+      );
     });
 
     it("destroy works", function () {